+++ CHANGELOG.md
@@ -10,13 +10,16 @@ Swift 3.0.0 is now the minimum Swift version supported.
 
 ### Enhancements
 
-* None.
+* Add change notifications for individual objects with an API similar to that
+  of collection notifications.
 
 ### Bugfixes
 
 * Fix Realm Objective-C compilation errors with Xcode 8.3 beta 1.
 * Fix several error handling issues when renewing expired authentication
   tokens for synchronized Realms.
+* Fix a race condition leading to bad_version exceptions being thrown in
+  Realm's background worker thread.
 
 2.3.0 Release notes (2017-01-19)
 =============================================================